Add $shown_ids to ids to be excluded from LMP query
The javascript code incrementing the page count has been moved to after the AJAX call. Previously, the posts that are included in $shown_ids were on page 0 of the results of the LMP query. They are now not included in the query return, so the "first" page of results is page 0. `paged` cannot be > 0 before the query, so the increment function has moved to after the query. I changed this in the minified js as well.