Undoing javascript fix since it wouldn't work on !home pages; fixing in largo_load_more_posts instead. Also bonus fix for determining is_home
Javascript fix from 08dfe5373f268fe9a7730e0d54a4c6909cad705d worked on homepages, but not on not-homepage archive pages. Their page 0 posts were actually in page 0, and not in $shown_ids.

This fix undoes the changes to the javascript, and instead fixes the homepage `paged` count in `largo_load_more_posts`, where there is already logic to determine if the query is on the home page.