Fix LMP dom on series landing pages.
Landing pages were loading partials/content when
partials/content-series is used on landing pages.

 - [ ] added an indicator sent with LMP data to alert the dom
generating function that this is a series landing page
 - [ ] pass $opt variable that partials/content-series uses to
determine the dom to return.
 - [ ] be a good person and restore global $post after the landing page
is done using it.

We can probably (in the future) overhaul the series page template to
simplify how it asks and controls dom.