Add the slug meta box.
In Largo 0.3, we started to hide the slug box from the post screen.

The problem with that is, this meta box contains the <input> field that
holds the post_name slug value. If the box is removed, the value from
the permalink editor is never POSTed and never saved to the database.

Reported by Wisconsin Watch.