Apply a filter to the partial loaded on the search page and in LMP.
This will not break Chicago Catalyst, which has both the 'content-archive' and 'content-search' partials.
Chicago Catalyst is currently broken.

Chicago Catalyst currently loads the 'content-search' partial in search.php and loads the 'content-archive'
partial in LMP. Chicago Catalyst and The Lens NOLA are the only sites defining content-search, and The Lens
is not on Largo 0.current, and thus is not affected by this pull request.

We should load the search partial in LMP on search pages, or not at all.