Probable fix for https://github.com/INN/Largo/issues/1381
I'm not able to duplicate the bug described in https://github.com/INN/Largo/issues/1381, but this looks like a fix: exclude the current post from the posts that get returned in any query in the Largo_Related class, which is what feeds the Largo Related Posts widget.

Please watch tests on this one.